// packages/common-dto/src/error/model/CustomErrorType.ts
export enum CustomErrorType {
  AUTHENTICATION_FAILED = 'AUTHENTICATION_FAILED',
  BAD_REQUEST = 'BAD_REQUEST',
  BAD_STATE = 'BAD_STATE',
  CANCELLED = 'CANCELLED',
  CONFLICT = 'CONFLICT',
  EXPIRED = 'EXPIRED',
  FORBIDDEN = 'FORBIDDEN',
  GONE = 'GONE',
  HTTP_COMMUNICATION = 'HTTP_COMMUNICATION',
  IM_A_TEAPOT = 'IM_A_TEAPOT',
  INVALID_CONFIGURATION = 'INVALID_CONFIGURATION',
  MALFORMED = 'MALFORMED',
  METHOD_NOT_ALLOWED = 'METHOD_NOT_ALLOWED',
  NETWORK_ERROR = 'NETWORK_ERROR',
  NOT_ACCEPTABLE = 'NOT_ACCEPTABLE',
  NOT_FOUND = 'NOT_FOUND',
  NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
  REQUEST_TIMEOUT = 'REQUEST_TIMEOUT',
  SERVER_ERROR = 'SERVER_ERROR',
  TIMEOUT = 'TIMEOUT',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  UNAUTHORIZED = 'UNAUTHORIZED',
  UNHANDLED_CLIENT_ERROR = 'UNHANDLED_CLIENT_ERROR',
  USER_NOT_ALLOWED = 'USER_NOT_ALLOWED',
}